Make test fail
I added an expectation where a unit in a foreign province can NOT make
use of the railroad rule. This means that the railroad rule only applies
when a unit originates from within their own nation's province and it
does not apply when the unit is entering a home province from outside.

I tried to fix the implementation to account for this but I was
unfortunately unable to understand the code and any changes I made
seemed to result in unexpected failures/outcomes.