Add convoy
There is an additional rule, which is that a single fleet can only
convoy one army per turn. I think we should implement this rule
elsewhere since it doesn't seem to have bearing on the "neighbors for"
an army. I guess we would put this limitation in at a different place
in the stack. What do you think, @jreut?