Reorganize code into Neighbors class
I'm really not sure if this is an improvement or just shuffling bad code
from one place to another. What do you think, @jreut? I think what I may
be looking for here is a pipeline or maybe something else? I do think
it's nice to separate and name the functions that *add* destinations as
opposed to those that *remove* destinations.