Use Go server and WebSockets
This commit makes the front end rely on the Go server, which will
connect two players over WebSockets. Currently, it only works with two
players. This commit also includes material for deploying the Go server
on EC2. Lastly, it adds to the Go server the ability to handle game
logs, thus making the game somewhat playable.